+++ src/dse_polars/__init__.py
@@ -6,7 +6,7 @@ except PackageNotFoundError:
     # Package is not installed (e.g., running from a local script)
     __version__ = "unknown"
 
-from .dse import DSE, drop_subref
+from .dse import DSE
 from .texts import DSEPassages, ctsurn_contains
 from .images import CitableIIIFService, urn2info_url, info_url2urn, roi, strip_roi
 from .urnutils import passagecomponent_re
@@ -14,7 +14,7 @@ from .urnutils import passagecomponent_re
 __all__ = ["DSE", 
            "DSEPassages",
            "ctsurn_contains",
-           "drop_subref",  
+         
                  
            "CitableIIIFService",
            "urn2info_url",